Add files via upload
Based on Deep Credit Risk


http://www.deepcreditrisk.com/data--code-private.html


The data set is in panel form and reports origination and performance observations for 50,000 residential U.S. mortgage borrowers over 60 periods. For training pruposes we work with a 10% sample (the full sample can be downloaded here). The periods have been deidentified. As in the real world, loans may originate before the start of the observation period (this is an issue where loans are transferred between banks and investors as in securitization). The loan observations may thus be censored as the loans mature or borrowers refinance. The data set is a randomized selection of mortgage-loan-level data collected from the portfolios underlying U.S. residential mortgage-backed securities (RMBS) securitization portfolios and provided by International Financial Research (www.internationalfinancialresearch.org).